Funcionalidades completas

+++ frontend/src/components/Header.jsx
@@ -4,6 +4,8 @@ const Header = () => {
   // Inicializar el estado 'usuario' basado en el localStorage
   const [usuario, setUsuario] = useState(() => sessionStorage.getItem("usuario"));
 
+  const tokenUsuario = sessionStorage.getItem("token");
+
   useEffect(() => {
     const idUsuario = sessionStorage.getItem("idUsuario");
     if (idUsuario) {
@@ -12,9 +14,31 @@ const Header = () => {
   }, []); // Solo se ejecuta una vez al montar el componente
 
   const cerrarSesion = () => {
+
+    const logout = async () => {
+      try {
+        const response = await fetch("http://localhost:8000/api/logout", {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+            "Authorization": `Bearer ${tokenUsuario}`,
+          },
+        });
+  
+        const data = await response.json();
+  
+        if (response.ok) {
+          console.log("Logout exitoso:", data.message);
+          sessionStorage.clear();
+          window.location.href = "/";
+        } 
+      } catch (e) {
+        console.log(e.message);
+      }
+    };
+
+    logout();
     // Limpiar el localStorage y redirigir al login
-    sessionStorage.clear();
-    window.location.href = "/";
   };
 
   return (
